+++ beacon-chain/blockchain/fork_choice_test.go
@@ -256,32 +256,3 @@ func TestVoteCount_IncreaseCountCorrectly(t *testing.T) {
 		t.Errorf("Expected 1 vote, received %d", count)
 	}
 }
-
-func TestBlockChildren(t *testing.T) {
-	genesisBlock := b.NewGenesisBlock([]byte{})
-	genesisHash, err := hashutil.HashBeaconBlock(genesisBlock)
-	if err != nil {
-		t.Fatal(err)
-	}
-	targets := []*pb.BeaconBlock{
-		{
-			Slot:             9,
-			ParentRootHash32: genesisHash[:],
-		},
-		{
-			Slot:             5,
-			ParentRootHash32: []byte{},
-		},
-		{
-			Slot:             8,
-			ParentRootHash32: genesisHash[:],
-		},
-	}
-	children, err := BlockChildren(genesisBlock, targets)
-	if err != nil {
-		t.Fatalf("Could not fetch block children: %v", err)
-	}
-	if len(children) != 2 {
-		t.Errorf("Expected %d children, received %d", 2, len(children))
-	}
-}
